requirements.txt
flask
python-docx
requests
